// client/models/Account.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type Account = {
    name: string;
    accountId: string;
    mask: string;
    officialName: string;
    type: string;
    subtype: string;
    balances: number;
    groupId: number;
    itemId: number;
    id: number;
};